wizard: migrate image step to seperate file
The create image upload wizard should be split into multiple files. This
improves readability and makes it simpler to refactor and expand the
wizard and its steps. The image step is now in its own seperate
component.
